💥 Massive refactor of codebase
This update introduces a stupidly large amount of changes changing the entire architecture of the site to decrease the amount of code needed to implement new features as well as make lots of improvement to security and efficiency.

The update splits the codebase into three main components a common PHP framework built from the ground up using the PHP & MySQL: Novice to Ninja - 6th Edition book to use on any future sites, a second framework specific to this particular site and, finally, the templates, assets and the single entry point (index.php).

Put together these form the entire site. This commit also implements several of the client's features and will make the remaining features a breeze to implement.

- Update site to use a single entry point (/public/index.php)
- Introduce PSR-4 Autoloader
- Create two separate frameworks (site specific (/classes/JosJobs/) and non-site specific (/classes/CupOfPHP))
- CupOfPHP
    - Authentication class providing auth for the website
    - DatabaseTable class providing ORM and DB Abstraction Layer
    - EntryPoint (self explanatory) contains loadTemplate, checkUrl methods as well
    - IRoutes interface for defining route methods
- JosJobs
    - Controllers - Manage routes for particular areas of the site
        - Auth - for login, logout, 403 (forbidden) routes
        - Category - Create, Read, Update and Delete routes
        - Job - Create, Read, Read Privileged, Update, Delete, Apply and View Applications routes
        - Location - Create, Read, Update and Delete routes
        - Root - Homepage, About, FAQ and Dashboard routes
        - User - Create, Read, Update and Delete routes
    - Entities - Represent database records for ORM stuff
        - Category
        - Job
        - Location
        - User
    - JosJobsRoutes
        - Implements IRoutes, lists all the site routes, the permissions they require and what RESTful methods they support. E.g. GET and POST
        - Methods for authentication and layoutVars too
- Refine templates (merge create and update)
- Implement client's features (Closes #7 and #8)